Split identify interfaces into seperate files
